add int8/16/32 stressor and refactor cuda backend structure
